seperate song controls into its own component
