+++ frontend/src/components/app/song-controls/index.ts
@@ -1,0 +1,1 @@
+export * from './song-controls';
